Add script to list all cards with activity in date range
I noticed that some of the cards we'd spent time on during a given week
weren't appearing in the output of list-archived-trello-cards.rb. This
is because the script only finds archived cards where the
last_activity_date is within the given date range. I think what I'm
really interested in is seeing all cards that have had activity during
the date range, irrespective of whether that was the last activity, and
irrespective of whether they've been archived.