Only consider cards with actions during the date range
This is to work around some odd behaviour in the Trello API. It would
appear that the last_activity_date can be later than the date of any
actions associated with the card. This can lead to cards erroneously
being listed by this script.

There's a little more information in
https://gist.github.com/chrisroos/7899b2bc626e289c45ff63803416f206.