Improve script to identify neglected cards
I noticed that this script wasn't identifying cards that hadn't had any
visible activity (i.e. visible in the Trello UI) within the last two
weeks.

This is because `Card#last_activity_date` is updated for things we're
not particularly interested in, a change in position for example. I'd
hoped we could use the date of the latest `Card#actions` but actions
also include things like cards moving position so that's no use.

I think what we're really interested in is cards that have been created
more than n weeks ago and that haven't had any comments within the same
n weeks.